Add ideal gas test
Run HMC on subsets of an ideal gas system, where the subsets are selected based on a geometric criterion.

Assert that an observable based on this criterion doesn't drift after a large number of updates when using local_resampling_move, and assert that it does drift when using an ablated version of local_resampling_move.