Simplify test_setting_up_restraints_using_smarts
mol_a and mol_b have a large MCS (15 atoms including Hs, or 6 atoms excluding them), but test asserted that the core contained exactly 2 atoms. This is the result obtained using distance-based heavy-atom MCS on random conformers with seed = 814, but picking a different random seed causes the MCS to contain any number of atoms between 0 and 6.

This commit removes the test's dependence on random conformers.