Reduce memory usage in RBFE example script (#911)
* Reduce memory usage in RBFE example script

Previously, references to results for all edges were retained in the
list of futures. This change avoids maintaining references to futures
by moving the saving of results into the concurrently-executed
function.

Note: this slightly reduces the granularity of the concurrency; here
we parallelize over edges instead of over edges and solvent/complex as
we did previously.

* Remove redundant parameterization

* Revert seed offset removal to reduce diff

* Inline args

* Explicitly block until subprocesses finish