Minor optimization of barostat
Generates random values for barostat in batches to reduce overhead of
generating random values for the scaling and metropolis check.

Also refactors the potentials to be a bit more generic for running
number of blocks that are exactly N // tpb. Doesn't seem to impact
Hif2a/DHFR, but might have some utility as N gets large